// src/types/grpc/proto/server/auth.ts
/**
* This file is auto-generated by nestjs-proto-gen-ts
*/

import { Observable } from 'rxjs';
import { Metadata } from '@grpc/grpc-js';

export namespace auth {
    export interface AuthService {
        getUserByMobile(
            data: GetUserByMobileWalletRequest,
            metadata?: Metadata,
            ...rest: any[]
        ): Observable<GetUserByMobileResponse>;
        verifyToken(
            data: VerifyTokenRequest,
            metadata?: Metadata,
            ...rest: any[]
        ): Observable<VerifyTokenResponses>;
    }
    export interface VerifyTokenRequest {
        token?: string;
    }
    export interface VerifyTokenResponses {
        message?: string;
        data?: auth.User;
        status?: number;
    }
    export interface GetUserByMobileWalletRequest {
        mobile?: string;
    }
    export interface GetUserByMobileResponse {
        message?: string;
        data?: auth.User;
        status?: number;
    }
    export interface Role {
        id?: string;
        type?: string;
    }
    export interface User {
        id?: string;
        mobile?: string;
        roles?: auth.Role[];
        status?: string;
        pin?: string;
    }
}